requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

attrs==19.3.0
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
django-countries==5.5
django-dotenv==1.4.2
django-seed==0.2.2
django-storages==1.9.1
django==2.2.5
docker-compose==1.25.4
docker[ssh]==4.2.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
faker==4.0.1
future==0.16.0
idna==2.7
importlib-metadata==1.5.0 ; python_version < '3.8'
jmespath==0.9.5
jsonschema==3.2.0
paramiko==2.7.1
pathspec==0.5.9
pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
pynacl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
pyyaml==5.2
requests==2.20.1
semantic-version==2.5.0
sentry-sdk==0.14.1
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.2
urllib3==1.24.3 ; python_version != '3.4'
wcwidth==0.1.8
websocket-client==0.57.0
zipp==3.0.0
boto3==1.10.26